[BPK-4276] Restructure L and M component pages
