+++ docs/src/pages/ListsPage/WebListsPage.js
@@ -1,0 +1,61 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2016-2019 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+import { BpkList, BpkListItem } from 'bpk-component-list';
+import listReadme from 'bpk-component-list/README.md';
+
+import { WebComponentPage } from '../../components/ComponentPage';
+
+const components = [
+  {
+    id: 'default',
+    title: 'Default',
+    examples: [
+      <div>
+        <BpkList>
+          <BpkListItem>Apples</BpkListItem>
+          <BpkListItem>
+            Oranges
+            <BpkList>
+              <BpkListItem>Tangerines</BpkListItem>
+              <BpkListItem>Mandarins</BpkListItem>
+              <BpkListItem>Satsumas</BpkListItem>
+            </BpkList>
+          </BpkListItem>
+          <BpkListItem>Pears</BpkListItem>
+        </BpkList>
+        <BpkList ordered>
+          <BpkListItem>First</BpkListItem>
+          <BpkListItem>Second</BpkListItem>
+          <BpkListItem>Third</BpkListItem>
+        </BpkList>
+      </div>,
+    ],
+  },
+];
+
+const WebListPage = () => (
+  <WebComponentPage
+    readme={listReadme}
+    examples={components}
+    packageName="bpk-component-list"
+  />
+);
+
+export default WebListPage;
