+++ docs/src/pages/MapPage/NativeMapPage.js
