--- docs/src/pages/TappableLinkLabelPage/TappableLinkLabelPage.js
@@ -1,40 +1,0 @@
-/*
- * Backpack - Skyscanner's Design System
- *
- * Copyright 2016-2019 Skyscanner Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import React from 'react';
-
-import IntroBlurb from '../../components/IntroBlurb';
-import DocsPageWrapper from '../../components/DocsPageWrapper';
-import IOS from '../IOSTappableLinkLabelPage';
-
-const blurb = [
-  <IntroBlurb>
-    Tappable link labels provide a way to inclue links within a block Backpack
-    styled text.{' '}
-  </IntroBlurb>,
-];
-
-const TappableLinkLabelPage = () => (
-  <DocsPageWrapper
-    title="Tappable link label"
-    blurb={blurb}
-    iosSubpage={<IOS wrapped />}
-  />
-);
-
-export default TappableLinkLabelPage;
